pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = "Laboratorio 2 - Fundamentos Ingenieria Computacional"
authors = ["Luis Ballado <luis.ballado@cinvestav.mx>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
cryptography = "^38.0.1"
pycryptodome = "^3.15.0"
pycryptodomex = "^3.15.0"
matplotlib = "^3.6.0"
tabulate = "^0.8.10"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
